+++ packages/export-web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/export-web
 
+## 0.31.0-pre.1
+
+### Patch Changes
+
+- @glissade/backend-canvas2d@0.31.0-pre.1
+- @glissade/core@0.31.0-pre.1
+- @glissade/scene@0.31.0-pre.1
+
 ## 0.31.0-pre.0
 
 ### Patch Changes